requirements.txt
# Google Cloud dependencies
google-cloud-storage>=2.0.0
google-cloud-bigquery>=2.34.0
functions-framework>=3.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# Testing and linting
pytest>=6.2.5
pytest-cov>=2.12.1
flake8>=3.9.2
black>=21.6b0
